+++ Casino.bat
@@ -1,5 +1,5 @@
 @echo off
 chcp 65001 >nul
 del DeustoCasino.exe
-gcc main.c menus.c personas.c fichero.c TragaPerras.c dado.c music.c carrera.c logger.c config.c blackjack.c bd.c sqlite3.c -o DeustoCasino -lwinmm -mconsole
+gcc Menus/main.c Menus/menus.c Model/personas.c Persistencia/fichero.c Juegos/TragaPerras.c Juegos/dado.c Music/music.c Juegos/carrera.c Logger/logger.c Persistencia/config.c Juegos/blackjack.c bd.c Persistencia/sqlite3.c -o DeustoCasino -lwinmm -mconsole
 DeustoCasino.exe
