+++ Juegos/dado.h
@@ -1,6 +1,6 @@
 #ifndef DADO_H_
 #define DADO_H_
-#include "personas.h"
+#include "Model/personas.h"
 int inicializar_dado(int recipientesCantidad, int selecionado);
 void lanzar_dado(int recipientesCantidad,int ganador);
 int ganar(int victoria,int cantidadRecipiente);
