+++ Persistencia/bd.c
