+++ Persistencia/config.c
