+++ Persistencia/fichero.h
@@ -1,6 +1,6 @@
 #ifndef FICHERO_H_
 #define FICHERO_H_
-#include "personas.h"
+#include "Model/personas.h"
 
 
 void guardarNuevaPersona(Cliente cliente, char*fichero );
