--- menus.h
@@ -1,10 +1,0 @@
-#ifndef MENUS_H_
-#define MENUS_H_
-
-char menuPrincipal();
-char menuJuegos();
-char menuSesion();
-char menuTragaperras();
-char menuCarrera();
-
-#endif /* MENUS_H_ */
